Create README.md
In this project, I used Python, Jupyter Notebook, and the Pandas library to analyze user data for a video game company. I used I analyzed data of in-game purchases by players. I aggregated and analyze the data to view trends in the purchasing behavior of the players. These were the key insights I found in my analysis:
•	Most of the players were male.
•	The 20-24 age group made most of the purchases in the game.
•	It would be interesting to see the ages of the top five spenders to see if they are in the 20-24 age group that purchased the most items and had the highest purchase amount.
The first steps I took to complete this analysis was to import the Pandas and Numpy Python libraries, and then read in the JSON file that was the basis of my analysis into my Jupyter Notebook. I then converted the output into a data frame. I used basic Pandas methods and functions to do my analysis.